requirements.txt
Django==3.1.7
pymysql==1.0.2
django-admin-rangefilter==0.7.0
django-jazzmin==2.4.1
fcm-django==0.3.7
coreapi==2.3.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
django-sslserver==0.22
six==1.15.0
django-bootstrap4==2.3.1
django-bootstrap-datepicker-plus==3.0.5
WeasyPrint==52.4
django-daterangefilter==1.0.0
django-csp==3.7
django-chartjs==2.2.1
